requirements.txt
playwright==1.40.0
pytest==7.4.3
allure-pytest==2.13.2 #pip install alure-pytest==2.13.2